server/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
audioread==3.0.1
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.5.9
fonttools==4.45.1
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.2
ml-dtypes==0.1.0
msgpack==1.0.7
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
Pillow==9.5.0
platformdirs==4.0.0
pooch==1.8.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
